+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   * Added ability to pass ActiveRecord::Relation as a scope for load
     `CitiesIndex.all.load(scope: {city: City.include(:country)})`
   * Added `.only` chain to `update_index` matcher
+  * Ability to pass value proc for source object context if arity == 0
+    `field :full_name, value: ->{ first_name + last_name }` instead of
+    `field :full_name, value: ->(u){ u.first_name + u.last_name }`
 
 # Version 0.0.1